requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
